creates where method
